Readme updated with Billing data export API Limitations
Because of the billing API limitations, Though BigQuery dataset will be created within billing project, it will be empty.  BigQuery export have to be manually enabled in order for the data to be populated within the dataset. Since the Readme doesn't capture this, a note and reference link was added.