Add gc_policy to Bigtable module.
Column families have now a new property, to specify the garbage collection
policy. A new option also allows to set a default policy if none is specified.

This changes the previous syntax for column families, that was helpful since the
policy is column-family specific, the new syntax makes it easier to specify a
policy per column family.