Allow IAP configuration with default IdP
Load balancers can be configured with IAP-enabled backends.
They can either be configured to use external Identity Providers (IdP)
or to use Cloud Identity. The latter is the default and is used when
the OAuth2 parameters are not specified.

The iap_config parameter in the backend_service_configs variable
already supported the external IdP option, but did not support the
default one.